+++ Editor/Converter/FromUTS2/UniversalToonShaderInfo.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-namespace UnityEditor.Rendering.Toon
-{
-    public class UniversalToonShaderInfo
-    {
-    }
-}